Notebook with differnt other strategies to cluster
In general the idea was not to gets ome colinearity but to try to reduce teh neotm of the Hankel matrices

TODO: in some cases the algorithm does explode. Here it might be worththwile to check it. Maybee it has something to do with the noemalization strategy if Bs and Cs are involved